Created a CommandFactory and a first Command instance, to start migrating all the methods to their corresponding individual commands
